Moved all state to context.js
