+++ client/App.jsx
@@ -1,48 +1,12 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { StateProvider } from './Context';
 import QueriesContainer from './Containers/QueriesContainer'
 import Menu from './Components/Menu'
 
-// class App extends Component {
-//   render() {
-
-//     return (
-//       <div>
-//         <h1>Hello World</h1>
-//       </div >
-//     )
-//   }
-// }
-
 const App = () => {
-  const initialState = {
-    greeting: 'hello Sophie',
-  };
-  const reducer = (state, action) => {
-    switch (action.type) {
-      case 'newGreeting':
-        return {
-          ...state,
-          greeting: action.newGreeting
-        }
-      case 'addURL':
-        return {
-          ...state,
-          url: action.addURL
-        }
-      case 'submitEndpoint':
-        return {
-          ...state,
-          endpoint: action.submitEndpoint
-        }
-      default:
-        return state;
-    }
-  }
-
   return (
     <div>
-      <StateProvider initialState={initialState} reducer={reducer}>
+      <StateProvider>
         <QueriesContainer />
         <Menu />
       </StateProvider>
@@ -50,6 +14,4 @@ const App = () => {
   );
 }
 
-
 export default App;
-
