+++ client/index.js
@@ -2,7 +2,10 @@ import { render } from 'react-dom';
 import React from 'react';
 import App from './App.jsx';
 
-render(<App />, document.getElementById('root'));
+render(
+  <App />, 
+  document.getElementById('root')
+  );
 
 if (module.hot) {
   module.hot.accept();
